Refactor menu page to a server side component that fetches data, add page content in menuContent component
