+++ src/app/_components/menuContent/menuContent.module.scss
@@ -1,0 +1,104 @@
+@import '../../../styles/mantine';
+
+.container {
+  padding: 48px 0;
+  @media (max-width: $mantine-breakpoint-sm) {
+    padding: 0 0 48px 0;
+  }
+
+  h2 {
+    grid-column: 1 / -1;
+    justify-self: flex-start;
+  }
+
+  .filterTop {
+    @media (max-width: $mantine-breakpoint-sm) {
+      transition: top 0.3s ease-out;
+      align-items: unset;
+      justify-content: center;
+      position: fixed;
+      display: flex;
+      top: var(--header-height);
+      z-index: 90;
+      background-color: var(--mantine-color-white-3);
+      padding: 16px 0;
+      left: 0;
+      width: 100%;
+    }
+  }
+
+  .filterTop,
+  .hideTop {
+    grid-column: 1 / -1;
+    > div {
+      display: flex;
+      justify-content: space-between;
+      gap: 24px;
+      @media (max-width: $mantine-breakpoint-sm) {
+        flex-direction: column;
+        gap: 8px;
+
+        h2 {
+          font-size: 34px;
+        }
+      }
+    }
+  }
+
+  @media (max-width: $mantine-breakpoint-sm) {
+    flex-direction: column;
+    align-items: unset;
+    gap: 8px;
+  }
+}
+
+.hideTop {
+  top: -100%;
+}
+
+.grid {
+  display: grid;
+  grid-template-columns: repeat(3, 328px);
+  gap: 32px 24px;
+  justify-content: center;
+
+  @media (max-width: $mantine-breakpoint-lg) {
+    grid-template-columns: repeat(2, 328px);
+  }
+
+  @media (max-width: $mantine-breakpoint-sm) {
+    grid-template-columns: repeat(1, 328px);
+    margin-top: 250px;
+  }
+}
+
+.error {
+  max-width: 600px;
+  width: 100vw;
+  padding: 48px 0;
+
+  @media (max-width: $mantine-breakpoint-sm) {
+    padding: 24px 0;
+    max-width: 100%;
+    grid-template-columns: 1fr;
+  }
+}
+
+.dropdown {
+  border-top-right-radius: 0;
+  border-top-left-radius: 0;
+  border-color: #4e4e4e6b;
+  border-top: 0;
+  z-index: 99;
+}
+
+.input {
+  border-radius: var(--mantine-radius-primary);
+  max-width: 400px;
+  width: 328px;
+  &[data-expanded] {
+    border-color: #4e4e4e6b;
+    border-bottom-left-radius: 0;
+    border-bottom-right-radius: 0;
+  }
+}
